backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
google-generativeai
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart